split container and component
