+++ src/routes/admin/player/player-list.component.js
@@ -1,41 +1,27 @@
-import { connect } from 'react-redux'
 import React from 'react'
 import PropTypes from 'prop-types'
-import PlayerListItem from './player-list-item.component'
+import PlayerListItemContainer from './player-list-item.container'
+import RoundSelectorContainer from './round-selector.container'
 
-const createSortReference = (roundId, roundEntries) => {
-  const sortReference = roundEntries.filter(e => e.round_id === roundId)
-    .map(e => [e.player_id, e.sort_order])
-  return Object.assign(...sortReference.map(r => ({[r[0]]: r[1]})))
-}
-
-const mapStateToProps = state => {
-  const roundToSort = state.adminApp.roundToSort
-  const roundEntries = state.adminApp.round_entries
-  const draft = state.adminApp.draft
-  const sortReference = createSortReference(roundToSort.id, roundEntries)
-
-  const players = state.adminApp.players.sort((p1, p2) =>
-    sortReference[p1.id] > sortReference[p2.id] ? 1 : -1
-  ).map(p => {
-    if (draft === null) return p
-    if (draft.player.id !== p.id) return p
-    return {...p, isEditing: true}
-  })
-
-  return {players}
-}
-
-const PlayerList = ({players}) => (
+const PlayerList = ({players, pars, loading}) => (
   <div>
+    <h2>出場選手一覧</h2>
+    <div className='row mb-2'>
+      <div className='col-auto mr-auto'/>
+      <div className='col-auto'>
+        <RoundSelectorContainer/>
+      </div>
+    </div>
     {players.map(p =>
-      <PlayerListItem key={p.id} {...p} />
+      <PlayerListItemContainer key={p.id} pars={pars} loading={loading} {...p} />
     )}
   </div>
 )
 
 PlayerList.propTypes = {
   players: PropTypes.array.isRequired,
+  pars: PropTypes.array.isRequired,
+  loading: PropTypes.bool.isRequired,
 }
 
-export default connect(mapStateToProps)(PlayerList)
+export default PlayerList
