+++ src/routes/admin/player/round-selector.component.js
@@ -1,27 +1,17 @@
-import { connect } from 'react-redux'
 import React from 'react'
 import PropTypes from 'prop-types'
-import { changeRoundToSort } from './player.module'
 
 const NOT_CHANGE = 'NOT_CHANGE'
-
-const mapStateToProps = state => ({
-  rounds: state.adminApp.rounds,
-})
-
-const mapDispatchToProps = dispatch => ({
-  onChangeRound: (id) => {
-    if (id === NOT_CHANGE) return
-    dispatch(changeRoundToSort(id))
-  }
-})
 
 const RoundSelector = ({
                          rounds,
                          onChangeRound,
                        }) => (
   <select className='custom-select'
-          onChange={e => onChangeRound(e.target.value)}>
+          onChange={e => {
+            if (e.target.value === NOT_CHANGE) return
+            onChangeRound(e.target.value)
+          }}>
     <option value={NOT_CHANGE}>並び順</option>
     {rounds.map(r =>
       <option key={`RoundSelector-${r.id}`} value={r.id}>{r.title}</option>
@@ -34,4 +24,4 @@ RoundSelector.propTypes = {
   onChangeRound: PropTypes.func.isRequired,
 }
 
-export default connect(mapStateToProps, mapDispatchToProps)(RoundSelector)
+export default RoundSelector
