+++ src/routes/admin/player/round-selector.container.js
@@ -1,0 +1,15 @@
+import { connect } from 'react-redux'
+import { changeRoundToSort } from './player.module'
+import RoundSelector from './round-selector.component'
+
+const mapStateToProps = state => ({
+  rounds: state.adminApp.rounds,
+})
+
+const mapDispatchToProps = dispatch => ({
+  onChangeRound: (id) => {
+    dispatch(changeRoundToSort(id))
+  }
+})
+
+export default connect(mapStateToProps, mapDispatchToProps)(RoundSelector)
